Move Java source file into correct position and adjust module based on Pazuzus recommendations
