+++ .dockerignore
@@ -1,7 +1,8 @@
 .git
 node_modules
 ios
-android
+android/*
+!android/app/src/main/java/expo/modules/efwifi/ExpoEfwifiModule.java
 dist
 .expo
 .cache
