+++ modules/ef-wifi/expo-module.config.json
@@ -1,5 +1,5 @@
 {
   "platforms": ["apple", "android"],
   "apple": { "modules": ["EfWifiModule"] },
-  "android": { "modules": ["expo.modules.efwifi.EfWifiModule"] }
+  "android": { "modules": ["expo.modules.efwifi.ExpoEfWifiModule"] }
 }
