+++ src/components/wifi/WifiSetup.tsx
@@ -81,13 +81,13 @@ export function WifiSetup({
 
       setBusy(true)
       if (Platform.OS === 'android') {
-        await addEnterpriseNetwork({
+        addEnterpriseNetwork(
           ssid: WIFI_SSID,
           identity: creds.identity,
           password: creds.password,
-          anonymousIdentity: WIFI_ANONYMOUS_IDENTITY,
-          domainSuffixMatch: WIFI_DOMAIN_SUFFIX_MATCH,
-        })
+          anonymous_identity: WIFI_ANONYMOUS_IDENTITY,
+          suffix_match: WIFI_DOMAIN_SUFFIX_MATCH,
+        )
       } else {
         // iOS: public profiles open a pre-filled static .mobileconfig directly (one-tap install);
         // custom creds can't be static, so they open the prefilled onsite page (tap Download there).
